nut-holder-test: Allow the hole to cut through the bottom
This is for longer screws that want to go further through the nut... Also this
increases the thickness of the nuts (and makes it adjustable) as the nylon nuts
I have are actually 2.6mm.

Text was also adjusted by an extra mm but it didn't seem to make a difference in
the end result on the Form 3.