power-module: Add dovetail attachers to the connector housings
Not 100% happy with this as they seem like they're too flimsy since they're only
2mm thick, but I'll do a test print, soon.

This also removes mirror_copy because it turns out it already exists in
BOSL (great minds, I suppose, but it's a handy function), and my verison without
offset causes some strange errors thanks to how everything is included (which is
a must to use a lot of the variables such as $tags).

There is still some diff goofiness that will get fixed in later versions of
openscad (namely diffing against an empty object is weird), which interferes
with the tag usage.

For the CTP2 module, I increased the height to make it look a bit nicer, and
give the wires a bit more clearance. 40mm will likely be the standard height.