Initial commit of the PCB holder
This is a module designed to slot into the power module, and allow adjustments
in x, y, and z, in order to position the USB ports with the opening, as it seems
tolerances on the boards I have for these components isn't high.

It's a lot of overengineering for these $5 parts, but hey if you want something
nice.

In theory I also only have to do this once.