pcb-holder: First test print
Main things are the parts now get rotated in their desired print orientation
when selecting individual parts, and of course addBase is applied to them.

addBase was funky in that it couldn't generate the projection for the bottom of
the clamp, lowering it by an epsilon value allowed it to work, I wonder if there
are planar issues there.