anderson-connectors: first draft of multi-holder
This adds a legs argument to the already complicated pp15_casing module, and
some refactoring that came from that.

I should check if the whole thing still fits with all these modifications.

All that's left is a few extra fillets and a test print (when the front piece is
ready).