terminal-block-cover: Make asymmetrical and print back piece
Where the back of the UMP-400-24 is the mains input. The screws are asymmetrical
there.

This also fixes an input error with the PSU width, as well as fixes a bug with
forgetting to account for the fact that I removed the bottom wall in the slot
measurement... it still barely fits though.