Fix problems with first print
Mainly, the connector wasn't given enough clearance (it needs 3mm).

This means that we need the overall thickness to be larger, and I don't
now how to feel about that. At this point it may be wise to just also
design the top part to hold down the keys, and possibly even an esp32
mount of some sort.