neokey1x4-holder: take3
So the bottom part didn't adhere the entire time to the platform (it
came off later on). I cleaned the platform this time, this may have been
old resin that caused it, but who knows. There wasn't any elephant's
foot either, strangely, which probably just means it did *not* do a good
job on the first layer on that part.

This also adds a slope to the nut hole to help it not sag as much, as it
will be printed in different layers.