take5
I printed take4 at 50um, but I think that makes the nut socket a lot
tighter.

I also realized it's impossible to put two modules together, so we need
nut sockets for both the top and the bottom, and just alternate.

The screw hole positions were also wrong and not relative to the PCB,
and the back clamp bit was too tall.

Also rounded out the standoffs to make it less sharp.