take8
Added back wire through holes, and add a bit more height to the top.

Sometimes capacitors are a bit taller due to soldering tolerances.

Still undecided whether or not the through holes are a good idea, as the
wire length has to be exact, and they have to be threaded through before
soldering.